+++ mobile/android/focus-android/app/src/main/java/org/mozilla/focus/input/InputToolbarIntegration.kt
@@ -18,12 +18,14 @@ import kotlinx.coroutines.flow.mapNotNull
 import mozilla.components.browser.domains.autocomplete.CustomDomainsProvider
 import mozilla.components.browser.domains.autocomplete.ShippedDomainsProvider
 import mozilla.components.browser.toolbar.BrowserToolbar
+import mozilla.components.browser.toolbar.display.DisplayToolbar.DisplayMargins
 import mozilla.components.compose.cfr.CFRPopup
 import mozilla.components.compose.cfr.CFRPopupProperties
 import mozilla.components.concept.toolbar.AutocompleteResult
 import mozilla.components.concept.toolbar.Toolbar
 import mozilla.components.lib.state.ext.flowScoped
 import mozilla.components.support.base.feature.LifecycleAwareFeature
+import mozilla.components.support.ktx.android.util.dpToPx
 import org.mozilla.focus.R
 import org.mozilla.focus.ext.components
 import org.mozilla.focus.ext.settings
@@ -118,6 +120,10 @@ class InputToolbarIntegration(
         )
 
         toolbar.display.setUrlBackground(urlBackground)
+        toolbar.display.setUrlBackgroundMargins(
+            DisplayMargins(8.dpToPx(toolbar.resources.displayMetrics), 0),
+        )
+
         toolbar.edit.setUrlBackground(urlBackground)
     }